finished up, mostly, with code cleanup

+++ application/views/articles/delete_results.php
@@ -12,19 +12,21 @@ if (!empty($articles)) { ?>
         <tbody>
         <?php foreach ($articles as $article): ?>
             <?php
-                $title = rawurlencode($article['title']);
-                echo "<tr class=\"search-list\" onclick=\"delete_article('{$title}', this)\">";
-            
+            if (strlen(trim($article['title'])) === 0 || strlen(trim($article['ajauthor'])) === 0 || strlen(trim($article['pubDate'])) === 0) {
+                continue;
+            }
+            $title = rawurlencode($article['title']);
+            echo "<tr class=\"search-list\" onclick=\"delete_article('{$title}', this)\">";
             ?>
-                <td>
-                    <?php echo $article['title']; ?>
-                </td>
-                <td>
-                    <?php echo $article['ajauthor']; ?>
-                </td>
-                <td>
-                    <?php echo $article['pubDate']; ?>
-                </td>
+            <td>
+                <?php echo $article['title']; ?>
+            </td>
+            <td>
+                <?php echo $article['ajauthor']; ?>
+            </td>
+            <td>
+                <?php echo $article['pubDate']; ?>
+            </td>
             </tr>
         <?php endforeach; ?>
         </tbody>
