+++ application/views/books/delete_results.php
@@ -11,15 +11,17 @@ if (!empty($books)) { ?>
         <tbody>
         <?php foreach ($books as $book): ?>
             <?php
+            if (strlen(trim($book['title'])) === 0 || strlen(trim($book['author'])) === 0) {
+                continue;
+            }
             echo "<tr class=\"search-list\" onclick=\"delete_book('{$book['isbn']}', this)\">";
-            
             ?>
-                <td>
-                    <?php echo $book['title']; ?>
-                </td>
-                <td>
-                    <?php echo $book['author']; ?>
-                </td>
+            <td>
+                <?php echo $book['title']; ?>
+            </td>
+            <td>
+                <?php echo $book['author']; ?>
+            </td>
             </tr>
         <?php endforeach; ?>
         </tbody>
